Major Refactor!
- Shift all async-tasks and file transfer code to WDM from LocalFileTransferActivity
- LocalFileTransferActivity only deals with setting up the UI now
